refactor(providers): tighten AiSession.prompt, migrate ports to _engine, resolve sessionId TODOs
- AiSession.prompt is now `string` — the branded Prompt is still accepted (subtype), but the three `as unknown as string` casts at the headless adapters' spawn / argv seams are gone.
- Migrate every port-shaped *Deps interface into the concept's _engine/: ClaudeProviderDeps / CopilotProviderDeps / CodexProviderDeps (headless), InteractiveClaudeDeps / InteractiveCopilotDeps / InteractiveCodexDeps + shared InteractiveSpawn, Claude/CopilotStreamLine/Parser + Usage/Envelope, SkillsAdapterFactoryDeps + CreateClaude/Copilot/CodexSkillsAdapterDeps, FileLogSink + FileLogSinkDeps, ProviderInstallGuidance / DetectInstalledProvidersOptions / InstallPlatform / WhichFn.
- Interactive sessionId: Claude and Copilot both expose a launch-time `--session-id <uuid>` override — the adapters now pre-generate a UUIDv7, pass it in, return it on success, and persist it to `session-id.txt` next to outputFile (mirrors the headless contract via persistSessionIdFile). Codex's interactive command has no equivalent override; the TODO is replaced with a documented constraint.
