+++ src/application/flows/settings-apply-preset/deps.ts
@@ -1,5 +1,5 @@
 import type { SettingsRepository } from '@src/domain/repository/settings/settings-repository.ts';
-import type { DetectInstalledProvidersOptions } from '@src/integration/system/detect-cli.ts';
+import type { DetectInstalledProvidersOptions } from '@src/integration/system/_engine/detect-cli.ts';
 import type { AiProvider } from '@src/domain/entity/settings.ts';
 
 export interface SettingsApplyPresetDeps {
