+++ src/integration/ai/providers/codex/interactive.ts
@@ -5,7 +5,8 @@ import type {
   InteractiveAiProvider,
   InteractiveAiProviderInput,
 } from '@src/integration/ai/providers/_engine/interactive-ai-provider.ts';
-import type { EventBus } from '@src/business/observability/event-bus.ts';
+import type { InteractiveCodexDeps } from '@src/integration/ai/providers/_engine/codex-interactive-deps.ts';
+import type { InteractiveSpawn } from '@src/integration/ai/providers/_engine/interactive-spawn.ts';
 import { InvalidStateError } from '@src/domain/value/error/invalid-state-error.ts';
 import { StorageError } from '@src/domain/value/error/storage-error.ts';
 import { IsoTimestamp } from '@src/domain/value/iso-timestamp.ts';
@@ -34,21 +35,6 @@ import { isCodexModel } from '@src/domain/value/settings-models/codex.ts';
  *
  * Docs: https://developers.openai.com/codex/cli/reference (top-level `codex` flags).
  */
-
-/** Test seam — same shape as `node:child_process.spawn` with `stdio: 'inherit'`. */
-export type InteractiveSpawn = (
-  command: string,
-  args: readonly string[],
-  options: { readonly stdio: 'inherit'; readonly cwd: string }
-) => ChildProcess;
-
-export interface InteractiveCodexDeps {
-  readonly eventBus: EventBus;
-  /** Test seam: defaults to `node:child_process.spawn`. */
-  readonly spawn?: InteractiveSpawn;
-  /** Override the shell name for tests / packaging. Defaults to `'bash'`. */
-  readonly command?: string;
-}
 
 const defaultSpawn: InteractiveSpawn = (command, args, options) =>
   nodeSpawn(command, [...args], { stdio: options.stdio, cwd: options.cwd });
@@ -138,10 +124,14 @@ export const createInteractiveCodexProvider = (deps: InteractiveCodexDeps): Inte
         at: IsoTimestamp.now(),
       });
 
-      // TODO(sessionId): see claude/interactive.ts for the rationale — stdio-inherit means the
-      // parent can't read stdout while the user owns the terminal, so the optional `sessionId`
-      // result field stays empty until a PTY-mirror or ~/.codex/sessions/ logfile probe lands.
-      // Per requirements REQ-3, absence is non-fatal.
+      // Codex's top-level (interactive) command does not accept a harness-supplied session id at
+      // launch — its only `--session-id` lives on `resume` / `fork` subcommands and is treated as
+      // a lookup key for an existing session, not a "use this UUID for the new conversation"
+      // override. Claude (`--session-id <uuid>`) and Copilot (`--session-id=<uuid>`) both ship
+      // that override and use it via the headless `session-id.txt` sidechannel; Codex
+      // intentionally leaves `sessionId` unset on success. Per the InteractiveAiProvider port
+      // contract, absence is non-fatal — subscribers that need a session correlation key fall
+      // back to the runner's session id from `AsyncLocalStorage`.
       if (exitCode === 0) return Result.ok({});
       return Result.error(
         new InvalidStateError({
