+++ src/integration/ai/skills/claude/adapter.ts
@@ -8,9 +8,9 @@
  * skills adapters, which only differ in `parentDir` and the convention text.
  */
 
-import type { Logger } from '@src/business/observability/logger.ts';
 import { createFilesystemSkillsAdapter } from '@src/integration/ai/skills/_engine/filesystem-skills-adapter.ts';
 import type { SkillsAdapter } from '@src/integration/ai/skills/_engine/skills-port.ts';
+import type { CreateClaudeSkillsAdapterDeps } from '@src/integration/ai/skills/_engine/claude-skills-adapter-deps.ts';
 
 const CONVENTION = [
   'Skills live under `.claude/skills/<name>/SKILL.md` in this repository. Each `SKILL.md`',
@@ -18,10 +18,6 @@ const CONVENTION = [
   'body. Before drafting, list `.claude/skills/` and read the `SKILL.md` of any folder',
   'whose `name` or `description` hints at sprint setup or post-task verification.',
 ].join(' ');
-
-export interface CreateClaudeSkillsAdapterDeps {
-  readonly logger?: Logger;
-}
 
 export const createClaudeSkillsAdapter = (deps: CreateClaudeSkillsAdapterDeps = {}): SkillsAdapter =>
   createFilesystemSkillsAdapter({
