+++ src/integration/ai/skills/codex/adapter.ts
@@ -7,9 +7,9 @@
  * {@link createFilesystemSkillsAdapter} — shared with the claude and copilot variants.
  */
 
-import type { Logger } from '@src/business/observability/logger.ts';
 import { createFilesystemSkillsAdapter } from '@src/integration/ai/skills/_engine/filesystem-skills-adapter.ts';
 import type { SkillsAdapter } from '@src/integration/ai/skills/_engine/skills-port.ts';
+import type { CreateCodexSkillsAdapterDeps } from '@src/integration/ai/skills/_engine/codex-skills-adapter-deps.ts';
 
 const CONVENTION = [
   'Skills live under `.agents/skills/<name>/SKILL.md` in this repository. Each `SKILL.md`',
@@ -17,10 +17,6 @@ const CONVENTION = [
   'body. Before drafting, list `.agents/skills/` and read the `SKILL.md` of any folder',
   'whose `name` or `description` hints at sprint setup or post-task verification.',
 ].join(' ');
-
-export interface CreateCodexSkillsAdapterDeps {
-  readonly logger?: Logger;
-}
 
 export const createCodexSkillsAdapter = (deps: CreateCodexSkillsAdapterDeps = {}): SkillsAdapter =>
   createFilesystemSkillsAdapter({
