+++ src/integration/ai/skills/copilot/adapter.ts
@@ -11,9 +11,9 @@
  * is the canonical Copilot-native location and what the CLI's docs steer authors toward.
  */
 
-import type { Logger } from '@src/business/observability/logger.ts';
 import { createFilesystemSkillsAdapter } from '@src/integration/ai/skills/_engine/filesystem-skills-adapter.ts';
 import type { SkillsAdapter } from '@src/integration/ai/skills/_engine/skills-port.ts';
+import type { CreateCopilotSkillsAdapterDeps } from '@src/integration/ai/skills/_engine/copilot-skills-adapter-deps.ts';
 
 const CONVENTION = [
   'Skills live under `.github/skills/<name>/SKILL.md` in this repository. Each `SKILL.md`',
@@ -21,10 +21,6 @@ const CONVENTION = [
   'body. Before drafting, list `.github/skills/` and read the `SKILL.md` of any folder',
   'whose `name` or `description` hints at sprint setup or post-task verification.',
 ].join(' ');
-
-export interface CreateCopilotSkillsAdapterDeps {
-  readonly logger?: Logger;
-}
 
 export const createCopilotSkillsAdapter = (deps: CreateCopilotSkillsAdapterDeps = {}): SkillsAdapter =>
   createFilesystemSkillsAdapter({
