+++ tests/integration/ai/providers/codex/interactive.test.ts
@@ -4,10 +4,8 @@ import type { ChildProcess } from 'node:child_process';
 import { absolutePath } from '@tests/fixtures/domain.ts';
 import { createCapturingBus } from '@tests/fixtures/capturing-event-bus.ts';
 import { CODEX_MODELS } from '@src/domain/value/settings-models/codex.ts';
-import {
-  createInteractiveCodexProvider,
-  type InteractiveSpawn,
-} from '@src/integration/ai/providers/codex/interactive.ts';
+import { createInteractiveCodexProvider } from '@src/integration/ai/providers/codex/interactive.ts';
+import type { InteractiveSpawn } from '@src/integration/ai/providers/_engine/interactive-spawn.ts';
 
 interface CapturingSpawnState {
   readonly spawn: InteractiveSpawn;
