+++ tests/unit/integration/ai/providers/copilot/headless-body-join.test.ts
@@ -1,8 +1,6 @@
 import { describe, expect, it } from 'vitest';
-import {
-  createCopilotStreamParser,
-  type CopilotStreamLine,
-} from '@src/integration/ai/providers/copilot/parse-stream.ts';
+import { createCopilotStreamParser } from '@src/integration/ai/providers/copilot/parse-stream.ts';
+import type { CopilotStreamLine } from '@src/integration/ai/providers/_engine/copilot-stream.ts';
 
 /**
  * Pins the body-accumulation contract used inside `copilot/headless.ts#spawnAttempt`:
